+++ vaadin-menu-bar.d.ts
@@ -1,2 +1,2 @@
 export * from './src/vaadin-menu-bar.js';
-export * from './@types/interfaces';
+export * from './src/interfaces';
